feat(env): Add production environment configuration
This commit introduces the configuration for the production environment.

An `.env.production` file has been created containing the base URL for the production API (`https://painel.softmed24h.com`). This file has also been added to the asset bundle in `pubspec.yaml` to ensure it is included in web builds.